React Component Task 3. Reusable comments & specialization
Containment

Create a new component named BodySection. The component does not know its children. It should output the following:

A div with the class bodySection
Within the div, a h2 element containing a title passed as a prop
Under the h2 the children of BodySection
Tips:

Rendering the following

<BodySection title="test">
  <p>test</p>
</BodySection>
Should generate:

<div className="bodySection">
  <h2>test</h2>
  <p>test</p>
</div>