React Component Task 5. Use the new components
in task_3/dashboard/src/App/App.js, modify the App component:

Wrap the CourseList component with the newly created BodySectionWithMarginBottom component. The title should be Course list
Wrap the Login component with the newly created BodySectionWithMarginBottom component. The title should be Log in to continue
Using the BodySection component, add a new block with the title News from the School. The component should contain a paragraph with some random text