React Redux Connectors and Providers Task 10. Connect Logout action creator to the Header
Modify the file Header/Header.js

Create a mapStateToProps function
Map the user props to the user within the Redux state
Connect the Header component to the function you created
Connect the Header component to the logout action creator
Modify the render function and remove any use of the Context. Instead use the user prop. When the user clicks on the link, it should now dispatch the logout action creator